prov/lnx: cleanup and reorganization
This patch doesn't include any functional changes.
Just cleans up the code in various ways:

- rename lnx_ops.c to lnx_srx.c to include only shared receive code
- move tag ops to lnx_msg.c to prepare for adding more functionality
while not having lnx_ops.c get too out of hand
- reduce extern functions declared in lnx.h/add static to appropriate
functions
- remove unused functions and definitions
- add missing definition of LNX_SUB_ID_BITS instead of hardcoded value
- standardize function declaration formatting
- fix formatting changes (line length, alignment, trailing whitespace)
- remove a few unnecessary comments
- cleanup unnecessary headers
