+++ prov/lnx/src/lnx_init.c
@@ -1,5 +1,6 @@
 /*
  * Copyright (c) 2022 ORNL. All rights reserved.
+ * Copyright (c) Intel Corporation. All rights reserved.
  *
  * This software is available to you under a choice of one of two
  * licenses.  You may choose to be licensed under the terms of the GNU
@@ -30,24 +31,9 @@
  * SOFTWARE.
  */
 
-#include "config.h"
-
-#include <assert.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <dirent.h>
-#include <ctype.h>
-
-#include <rdma/fi_errno.h>
-#include "ofi_util.h"
-#include "ofi.h"
-#include "ofi_str.h"
-#include "ofi_prov.h"
-#include "ofi_perf.h"
-#include "ofi_hmem.h"
-#include "rdma/fi_ext.h"
 #include "lnx.h"
+#include "ofi_prov.h"
+#include <ctype.h>
 
 #define LNX_PASSTHRU_TX_OP_FLAGS	(FI_INJECT_COMPLETE | \
 					 FI_TRANSMIT_COMPLETE | \
@@ -70,7 +56,7 @@ struct fi_tx_attr lnx_tx_attr = {
 	.inject_size 	= SIZE_MAX,
 	.size 		= SIZE_MAX,
 	.iov_limit 	= LNX_IOV_LIMIT,
-	.rma_iov_limit = LNX_IOV_LIMIT,
+	.rma_iov_limit	= LNX_IOV_LIMIT,
 };
 
 struct fi_rx_attr lnx_rx_attr = {
@@ -92,7 +78,7 @@ struct fi_ep_attr lnx_ep_attr = {
 	.max_order_raw_size 	= SIZE_MAX,
 	.max_order_war_size 	= SIZE_MAX,
 	.max_order_waw_size 	= SIZE_MAX,
-	.mem_tag_format = FI_TAG_GENERIC,
+	.mem_tag_format		= FI_TAG_GENERIC,
 	.tx_ctx_cnt 		= SIZE_MAX,
 	.rx_ctx_cnt 		= SIZE_MAX,
 	.auth_key		= NULL,
@@ -139,6 +125,32 @@ struct fi_info lnx_info = {
 	.fabric_attr = &lnx_fabric_attr
 };
 
+static int lnx_fabric_close(struct fid *fid)
+{
+	int rc, i, frc = 0;
+	struct lnx_core_fabric *cf;
+	struct lnx_fabric *lnx_fab;
+
+	lnx_fab = container_of(fid, struct lnx_fabric,
+			       lf_util_fabric.fabric_fid.fid);
+
+	for (i = 0; i < lnx_fab->lf_num_fabs; i++) {
+		cf = &lnx_fab->lf_core_fabrics[i];
+
+		rc = fi_close(&cf->cf_fabric->fid);
+		if (rc)
+			frc = rc;
+	}
+
+	rc = ofi_fabric_close(&lnx_fab->lf_util_fabric);
+	if (rc)
+		frc = rc;
+
+	free(lnx_fab->lf_core_fabrics);
+
+	return frc;
+}
+
 static struct fi_ops lnx_fabric_fi_ops = {
 	.size = sizeof(struct fi_ops),
 	.close = lnx_fabric_close,
@@ -162,7 +174,6 @@ static int lnx_wait_open(struct fid_fabric *fabric_fid,
 	}
 }
 
-
 static struct fi_ops_fabric lnx_fabric_ops = {
 	.size = sizeof(struct fi_ops_fabric),
 	.domain = lnx_domain_open,
@@ -175,22 +186,6 @@ static struct fi_ops_fabric lnx_fabric_ops = {
 	.trywait = fi_no_trywait
 };
 
-struct fi_provider lnx_prov = {
-	.name = OFI_LNX,
-	.version = OFI_VERSION_DEF_PROV,
-	.fi_version = OFI_VERSION_LATEST,
-	.getinfo = lnx_getinfo,
-	.fabric = lnx_fabric,
-	.cleanup = lnx_fini
-};
-
-struct util_prov lnx_util_prov = {
-	.prov = &lnx_prov,
-	.info = &lnx_info,
-	.flags = 0
-};
-
-/* this is a list of all possible links */
 struct dlist_entry lnx_links;
 
 struct lnx_fi_cache_entry {
@@ -224,15 +219,14 @@ static void lnx_trim(char *str)
 	str[i] = '\0';
 }
 
-
-void lnx_free_links(struct dlist_entry *links)
+static void lnx_free_links(struct dlist_entry *links)
 {
 	struct lnx_link_info *link;
 	struct lnx_fi_cache_entry *e;
 	struct dlist_entry *tmp, *tmp2;
 
-	dlist_foreach_container_safe(links, struct lnx_link_info, link,
-				     entry, tmp) {
+	dlist_foreach_container_safe(links, struct lnx_link_info, link, entry,
+				     tmp) {
 		dlist_foreach_container_safe(&link->link_providers,
 					     struct lnx_fi_cache_entry, e,
 					     entry, tmp2) {
@@ -245,8 +239,8 @@ void lnx_free_links(struct dlist_entry *links)
 	}
 }
 
-static int lnx_cache_info(struct dlist_entry *head,
-			  struct fi_info *info, bool new_prov)
+static int lnx_cache_info(struct dlist_entry *head, struct fi_info *info,
+			  bool new_prov)
 {
 	struct lnx_fi_cache_entry *e;
 	struct fi_info *itr;
@@ -287,13 +281,11 @@ insert:
 	return 0;
 }
 
-static struct lnx_link_info *
-lnx_get_link_by_dom(char *domain_name)
+static struct lnx_link_info *lnx_get_link_by_dom(char *domain_name)
 {
 	struct lnx_link_info *link;
 
-	dlist_foreach_container(&lnx_links, struct lnx_link_info, link,
-				entry) {
+	dlist_foreach_container(&lnx_links, struct lnx_link_info, link, entry) {
 		if (!strcmp(domain_name, link->fi_link->domain_attr->name))
 			return link;
 	}
@@ -301,12 +293,13 @@ lnx_get_link_by_dom(char *domain_name)
 	return NULL;
 }
 
-static int lnx_generate_link_info(struct fi_info **info, const struct fi_info *hints)
+static int lnx_generate_link_info(struct fi_info **info,
+				  const struct fi_info *hints)
 {
 	struct fi_info *itr, *fi = NULL, *next;
 	struct lnx_link_info *link;
 	struct lnx_fi_cache_entry *e;
-	size_t min_inject_size = SIZE_MAX, min_of_max_msg_size = SIZE_MAX;
+	size_t min_inject_size = SIZE_MAX, min_max_msg_size = SIZE_MAX;
 	size_t min_rx_size = SIZE_MAX, min_tx_size = SIZE_MAX;
 	size_t min_iov_limit = LNX_IOV_LIMIT;
 	int mr_mode = 0;
@@ -318,8 +311,7 @@ static int lnx_generate_link_info(struct fi_info **info, const struct fi_info *h
 	bool reset;
 	uint64_t caps = 0;
 
-	dlist_foreach_container(&lnx_links, struct lnx_link_info, link,
-				entry) {
+	dlist_foreach_container(&lnx_links, struct lnx_link_info, link, entry) {
 		link_name = calloc(sizeof(char), str_len);
 		if (!link_name)
 			return -FI_ENOMEM;
@@ -350,15 +342,16 @@ static int lnx_generate_link_info(struct fi_info **info, const struct fi_info *h
 
 			if (reset) {
 				caps = e->fi->caps;
-				min_inject_size = min_of_max_msg_size = SIZE_MAX;
+				min_inject_size = min_max_msg_size = SIZE_MAX;
 				min_rx_size = min_tx_size = SIZE_MAX;
 				min_iov_limit = LNX_IOV_LIMIT;
 				reset = false;
 			} else {
 				caps &= e->fi->caps;
 			}
-			if (e->fi->ep_attr->max_msg_size < min_of_max_msg_size)
-				min_of_max_msg_size = e->fi->ep_attr->max_msg_size;
+			if (e->fi->ep_attr->max_msg_size < min_max_msg_size)
+				min_max_msg_size =
+						e->fi->ep_attr->max_msg_size;
 			if (e->fi->tx_attr->inject_size < min_inject_size)
 				min_inject_size = e->fi->tx_attr->inject_size;
 			if (e->fi->tx_attr->iov_limit < min_iov_limit)
@@ -375,7 +368,8 @@ static int lnx_generate_link_info(struct fi_info **info, const struct fi_info *h
 						tmp[0] = '+';
 						tmp++;
 					}
-					memcpy(tmp, itr->domain_attr->name, len);
+					memcpy(tmp, itr->domain_attr->name,
+					       len);
 					tmp += (len - 1);
 				} else {
 					str_len = str_len * 2 + len;
@@ -389,13 +383,14 @@ static int lnx_generate_link_info(struct fi_info **info, const struct fi_info *h
 					tmp2 = link_name + str_len - 1;
 					tmp[0] = '+';
 					tmp++;
-					memcpy(tmp, itr->domain_attr->name, len);
+					memcpy(tmp, itr->domain_attr->name,
+					       len);
 					tmp += (len - 1);
 				}
 			}
 		}
 
-		link_name = realloc(link_name, strlen(link_name)+1);
+		link_name = realloc(link_name, strlen(link_name) + 1);
 		free(link->fi_link->fabric_attr->prov_name);
 		free(link->fi_link->fabric_attr->name);
 		free(link->fi_link->domain_attr->name);
@@ -410,17 +405,21 @@ static int lnx_generate_link_info(struct fi_info **info, const struct fi_info *h
 		next = fi_dupinfo(link->fi_link);
 		next->ep_attr->protocol = FI_PROTO_LNX;
 		next->domain_attr->mr_mode = mr_mode;
-		next->domain_attr->rx_ctx_cnt = lnx_util_prov.info->domain_attr->rx_ctx_cnt;
-		next->domain_attr->tx_ctx_cnt = lnx_util_prov.info->domain_attr->tx_ctx_cnt;
-		next->ep_attr->rx_ctx_cnt = lnx_util_prov.info->ep_attr->rx_ctx_cnt;
-		next->ep_attr->tx_ctx_cnt = lnx_util_prov.info->ep_attr->tx_ctx_cnt;
+		next->domain_attr->rx_ctx_cnt =
+				lnx_util_prov.info->domain_attr->rx_ctx_cnt;
+		next->domain_attr->tx_ctx_cnt =
+				lnx_util_prov.info->domain_attr->tx_ctx_cnt;
+		next->ep_attr->rx_ctx_cnt =
+				lnx_util_prov.info->ep_attr->rx_ctx_cnt;
+		next->ep_attr->tx_ctx_cnt =
+				lnx_util_prov.info->ep_attr->tx_ctx_cnt;
 		if (hints)
 			next->caps = caps & hints->caps;
 		else
 			next->caps = caps;
 		next->tx_attr->inject_size = min_inject_size;
 		next->tx_attr->iov_limit = min_iov_limit;
-		next->ep_attr->max_msg_size = min_of_max_msg_size;
+		next->ep_attr->max_msg_size = min_max_msg_size;
 		next->rx_attr->size = min_rx_size;
 		next->tx_attr->size = min_tx_size;
 
@@ -441,8 +440,7 @@ static int lnx_generate_link_info(struct fi_info **info, const struct fi_info *h
 	return FI_SUCCESS;
 }
 
-static struct fi_info *
-lnx_filter_info(struct fi_info *fi, char *domain)
+static struct fi_info *lnx_filter_info(struct fi_info *fi, char *domain)
 {
 	struct fi_info *itr, *tmp, *filtered_head = NULL, *filtered_tail = NULL;
 
@@ -478,9 +476,9 @@ lnx_filter_info(struct fi_info *fi, char *domain)
 	return (filtered_head) ? filtered_head : fi;
 }
 
-int lnx_getinfo_helper(uint32_t version, char *prov, char *domain,
-		       struct fi_info *lnx_hints, bool new_prov,
-		       struct dlist_entry *link)
+static int lnx_getinfo_helper(uint32_t version, char *prov, char *domain,
+			      struct fi_info *lnx_hints, bool new_prov,
+			      struct dlist_entry *link)
 {
 	int rc;
 	char *orig_prov_name = NULL, *orig_dom_name = NULL;
@@ -503,8 +501,8 @@ int lnx_getinfo_helper(uint32_t version, char *prov, char *domain,
 		lnx_hints->domain_attr->caps &= ~(FI_REMOTE_COMM);
 	}
 
-	rc = fi_getinfo(version, NULL, NULL, OFI_GETINFO_HIDDEN,
-			lnx_hints, &core_info);
+	rc = fi_getinfo(version, NULL, NULL, OFI_GETINFO_HIDDEN, lnx_hints,
+			&core_info);
 
 	lnx_hints->fabric_attr->prov_name = orig_prov_name;
 	lnx_hints->domain_attr->name = orig_dom_name;
@@ -521,9 +519,9 @@ int lnx_getinfo_helper(uint32_t version, char *prov, char *domain,
 	return rc;
 }
 
-int lnx_getinfo(uint32_t version, const char *node, const char *service,
-		uint64_t flags, const struct fi_info *hints,
-		struct fi_info **info)
+static int lnx_getinfo(uint32_t version, const char *node, const char *service,
+		       uint64_t flags, const struct fi_info *hints,
+		       struct fi_info **info)
 {
 	int rc;
 	bool new_prov;
@@ -547,8 +545,8 @@ int lnx_getinfo(uint32_t version, const char *node, const char *service,
 
 	if (strstr(linked_provs, "lnx")) {
 		FI_WARN(&lnx_prov, FI_LOG_FABRIC,
-			"Can't specify the lnx provider as part of the link: %s\n",
-			linked_provs);
+			"Can't specify the lnx provider as part of the link: "
+			"%s\n", linked_provs);
 		return -FI_EINVAL;
 	}
 
@@ -561,13 +559,10 @@ int lnx_getinfo(uint32_t version, const char *node, const char *service,
 	if (!lnx_hints)
 		return -FI_ENOMEM;
 
-	rc = ofi_exclude_prov_name(&lnx_hints->fabric_attr->prov_name, lnx_prov.name);
+	rc = ofi_exclude_prov_name(&lnx_hints->fabric_attr->prov_name,
+				   lnx_prov.name);
 	if (rc)
 		return rc;
-
-	/* get the providers which support peer functionality. These are
-	 * the only ones we can link*/
-	lnx_hints->caps |= FI_PEER;
 
 	/* Format:
 	 *  '+' appends another provider to the link.
@@ -614,9 +609,10 @@ int lnx_getinfo(uint32_t version, const char *node, const char *service,
 			domain = strtok_r(NULL, ",", &save_ptr2);
 			do {
 				lnx_trim(domain);
-				rc = lnx_getinfo_helper(version, provider, domain,
-						lnx_hints, new_prov,
-						&link->link_providers);
+				rc = lnx_getinfo_helper(version, provider,
+							domain, lnx_hints,
+							new_prov,
+							&link->link_providers);
 				if (rc)
 					return rc;
 				new_prov = false;
@@ -636,9 +632,8 @@ generate_info:
 	return rc;
 }
 
-static int
-lnx_setup_core_fab(struct lnx_fabric *lnx_fab, struct fi_info *info,
-		   void *context, int i)
+static int lnx_setup_core_fab(struct lnx_fabric *lnx_fab, struct fi_info *info,
+			      void *context, int i)
 {
 	int rc = -FI_ENOMEM;
 	struct lnx_core_fabric *cf;
@@ -658,9 +653,7 @@ fail:
 	return rc;
 }
 
-int
-lnx_setup_fabrics(char *name, struct lnx_fabric *lnx_fab,
-		  void *context)
+int lnx_setup_fabrics(char *name, struct lnx_fabric *lnx_fab, void *context)
 {
 	int rc;
 	struct lnx_link_info *link;
@@ -697,8 +690,8 @@ out:
 	return FI_SUCCESS;
 }
 
-int lnx_fabric(struct fi_fabric_attr *attr, struct fid_fabric **fabric,
-		void *context)
+static int lnx_fabric(struct fi_fabric_attr *attr, struct fid_fabric **fabric,
+		      void *context)
 {
 	struct lnx_fabric *lnx_fab;
 	int rc;
@@ -708,8 +701,8 @@ int lnx_fabric(struct fi_fabric_attr *attr, struct fid_fabric **fabric,
 		return -FI_ENOMEM;
 
 	rc = ofi_fabric_init(&lnx_prov, lnx_info.fabric_attr,
-			     lnx_info.fabric_attr,
-			     &lnx_fab->lf_util_fabric, context);
+			     lnx_info.fabric_attr, &lnx_fab->lf_util_fabric,
+			     context);
 	if (rc)
 		goto fail;
 
@@ -723,41 +716,31 @@ fail:
 	return rc;
 }
 
-void lnx_fini(void)
+static void lnx_fini(void)
 {
 	lnx_free_links(&lnx_links);
 	ofi_bufpool_destroy(global_recv_bp);
-}
-
-int lnx_fabric_close(struct fid *fid)
-{
-	int rc, i, frc = 0;
-	struct lnx_core_fabric *cf;
-	struct lnx_fabric *lnx_fab;
-
-	lnx_fab = container_of(fid, struct lnx_fabric, lf_util_fabric.fabric_fid.fid);
-
-	for (i = 0; i < lnx_fab->lf_num_fabs; i++) {
-		cf = &lnx_fab->lf_core_fabrics[i];
-
-		rc = fi_close(&cf->cf_fabric->fid);
-		if (rc)
-			frc = rc;
-	}
-
-	rc = ofi_fabric_close(&lnx_fab->lf_util_fabric);
-	if (rc)
-		frc = rc;
-
-	free(lnx_fab->lf_core_fabrics);
-
-	return frc;
 }
 
 void ofi_link_fini(void)
 {
 	lnx_prov.cleanup();
 }
+
+struct fi_provider lnx_prov = {
+	.name = OFI_LNX,
+	.version = OFI_VERSION_DEF_PROV,
+	.fi_version = OFI_VERSION_LATEST,
+	.getinfo = lnx_getinfo,
+	.fabric = lnx_fabric,
+	.cleanup = lnx_fini
+};
+
+struct util_prov lnx_util_prov = {
+	.prov = &lnx_prov,
+	.info = &lnx_info,
+	.flags = 0
+};
 
 LNX_INI
 {
